requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
requests
Pillow
python-dateutil>=2.8.0
numpy>=1.24.0
msal>=1.24.0
Office365-REST-Python-Client>=2.4.0
gspread>=5.10.0
google-auth>=2.20.0 